Updated Blog: BSc graduation and clean up
